requirements.txt
bottle==0.12.23
gitpython==3.1.27
bcrypt
python-dotenv
jwcrypto